+++ lib/functions/firestore.dart
@@ -1,0 +1,1 @@
+export '../src/functions/v2/providers/firestore.dart';