+++ readme/async_function.dart
@@ -1,10 +1,11 @@
 import 'dart:js_interop';
 import 'dart:js_interop_unsafe';
 
+import 'package:firebase_js_interop/admin/firestore.dart';
 import 'package:firebase_js_interop/functions.dart';
+import 'package:firebase_js_interop/functions/firestore.dart';
 import 'package:firebase_js_interop/js.dart';
 import 'package:firebase_js_interop/node.dart';
-import 'package:firebase_js_interop/admin.dart';
 
 void main() {
   exports['onMessageCreated'] = FirebaseFunctions.firestore.onDocumentCreated(