+++ readme/hello_world.dart
@@ -2,6 +2,7 @@ import 'dart:js_interop';
 import 'dart:js_interop_unsafe';
 
 import 'package:firebase_js_interop/functions.dart';
+import 'package:firebase_js_interop/functions/https.dart';
 import 'package:firebase_js_interop/node.dart';
 import 'package:firebase_js_interop/express.dart' as express;